api handles chart access authorization only, the rest is done by plugins (#155)
* api handles chart access authorization only, the rest is done by plugins

* bump orm to 3.14.2 and chart-core to 8.8.1

* load plugins before routes (so they can influence the Joi schema)

* only allow formats that are registered by plugins

* find first results with data
