+++ package.json
@@ -46,9 +46,9 @@
         "url": "git+https://github.com/datawrapper/datawrapper-api.git"
     },
     "dependencies": {
-        "@datawrapper/chart-core": "8.8.1",
+        "@datawrapper/chart-core": "^8.8.1",
         "@datawrapper/locales": "1.0.0",
-        "@datawrapper/orm": "^3.14.0",
+        "@datawrapper/orm": "^3.14.2",
         "@datawrapper/schemas": "^1.3.0",
         "@datawrapper/shared": "0.25.1",
         "@hapi/boom": "9.1.0",
