+++ src/routes/charts.js
@@ -5,7 +5,7 @@ const { Op } = require('@datawrapper/orm').db;
 const { decamelizeKeys, decamelize } = require('humps');
 const set = require('lodash/set');
 const mime = require('mime');
-const { Chart, ChartPublic, User, Folder, Plugin } = require('@datawrapper/orm/models');
+const { Chart, ChartPublic, User, Folder } = require('@datawrapper/orm/models');
 const { prepareChart } = require('../utils/index.js');
 const assignWithEmptyObjects = require('../utils/assignWithEmptyObjects');
 
@@ -235,9 +235,9 @@ function register(server, options) {
                         .required()
                         .description('5 character long chart ID.'),
                     format: Joi.string()
-                        .lowercase()
                         .required()
-                        .description('Export format (pdf, png, svg, zip)')
+                        .valid(...server.app.exportFormats.values())
+                        .description('Export format')
                 }),
                 payload: Joi.object({
                     unit: Joi.string().default('px'),
@@ -279,9 +279,9 @@ function register(server, options) {
                         .required()
                         .description('5 character long chart ID.'),
                     format: Joi.string()
-                        .lowercase()
                         .required()
-                        .description('Export format (pdf, png, svg, zip)')
+                        .valid(...server.app.exportFormats.values())
+                        .description('Export format')
                 }),
                 query: Joi.object({
                     unit: Joi.string().default('px'),
@@ -786,37 +786,40 @@ async function exportChart(request, h) {
     const { events, event } = server.app;
     const user = auth.artifacts;
 
-    const userPlugins = await user.getUserPluginCache();
-    const plugins = userPlugins && userPlugins.plugins ? userPlugins.plugins.split(',') : [];
-
-    if (params.format !== 'png' && !plugins.includes('export-pdf')) {
-        const pdfPlugin = await Plugin.findByPk('export-pdf');
-
-        if (pdfPlugin && pdfPlugin.is_private) {
-            return Boom.forbidden();
+    // authorize user
+    const chart = await Chart.findOne({
+        where: {
+            id: params.id,
+            deleted: { [Op.not]: true }
         }
-    }
+    });
 
-    if (params.format === 'zip' && !plugins.includes('export-zip')) {
-        const zipPlugin = await Plugin.findByPk('export-zip');
+    if (!chart) return Boom.notFound();
+    const mayEdit = await user.mayEditChart(chart);
+    if (!mayEdit) return Boom.notFound();
 
-        if (zipPlugin && zipPlugin.is_private) {
-            return Boom.forbidden();
-        }
-    }
+    // user is authorized to access chart
+    // further authoritzation is handled by plugins
 
     Object.assign(payload, params);
     try {
-        const result = await events.emit(
-            event.CHART_EXPORT,
-            {
-                data: payload,
-                userId: user.id,
-                auth,
-                logger
-            },
-            { filter: 'first' }
-        );
+        const result = (
+            await events.emit(
+                event.CHART_EXPORT,
+                {
+                    chart,
+                    user,
+                    data: payload,
+                    auth,
+                    logger
+                },
+                {
+                    filter: 'success'
+                }
+            )
+        ).find(res => res.data);
+
+        if (!result) return Boom.badImplementation();
 
         await request.server.methods.logAction(user.id, `chart/export/${params.format}`, params.id);
